pyproject.toml
[tool.poetry]
name = "naimium"
version = "0.1.0"
description = ""
authors = ["II9tka <fffxx@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
psycopg2-binary = "^2.9.6"
django-filter = "^23.1"
inflection = "^0.5.1"
pillow = "^9.5.0"
drf-yasg = "^1.21.5"
drf-nested-routers = "^0.93.4"
environs = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
